fix existing test cases

+++ lib/octobox/subjects/sync_subject_labels.rb
@@ -23,11 +23,13 @@ module Octobox
         }.map {
           |l| Label.new(github_id: l['id'], color: l['color'], name: l['name'], repository_id: repository.id)
         }
-        # insert new labels in bulk
-        Label.import new_labels_to_be_added
 
-        # create subject label mapping for New Added Labels
-        attach_label_to_subject(labels_to_be_added.collect(&:github_id))
+        if labels_to_be_added.present?
+          # insert new labels in bulk
+          Label.import labels_to_be_added
+          # create subject label mapping for New Added Labels
+          attach_label_to_subject(labels_to_be_added.collect(&:github_id))
+        end
 
         # update name and color of existing labels
         update_labels(
@@ -37,6 +39,8 @@ module Octobox
       end
 
       def update_labels(remote_labels, labels_for_update)
+        return if labels_for_update.blank?
+
         updated_labels = []
         remote_labels.each do |label|
           if matching_label = labels_for_update.select { |l| l.github_id == label['id'] }
@@ -52,14 +56,18 @@ module Octobox
       end
 
       def attach_label_to_subject(label_github_ids)
+        return if label_github_ids.blank?
+
         subject_labels = Label.where('github_id in (?)', label_github_ids).ids
         subject_labels.map!{ |label_id| SubjectLabel.new(subject_id: self.id, label_id: label_id) }
         SubjectLabel.import subject_labels
       end
 
       def detach_subject_labels_removed_on_remote(remote_label_ids)
+        return if remote_label_ids.blank?
+
         # filter all the labels removed from subject on remote and destroy
-        SubjectLabel.joins(:labels).where(
+        SubjectLabel.joins(:label).where(
           "subject_labels.subject_id = ? and labels.github_id not in (?)",
           self.id, remote_label_ids
         ).destroy_all
